--- a/src/Shared/Compatibility/SkipLocalsInitAttribute.cs
+++ b/src/Shared/Compatibility/SkipLocalsInitAttribute.cs
@@ -1,0 +1,15 @@
+namespace System.Runtime.CompilerServices;
+
+[AttributeUsage(
+      AttributeTargets.Module
+    | AttributeTargets.Class
+    | AttributeTargets.Struct
+    | AttributeTargets.Interface
+    | AttributeTargets.Constructor
+    | AttributeTargets.Method
+    | AttributeTargets.Property
+    | AttributeTargets.Event, Inherited = false)]
+internal sealed class SkipLocalsInitAttribute : Attribute
+{
+    public SkipLocalsInitAttribute() { }
+}