… src/Esprima/Properties/AssemblyInfo.cs