@adams85 2022-09-23
4-5% perf. gain for adding an attribute. What a catch! 👍